better home for error formatting
