+++ languageWorkbench/parserlib/pretty.ts
@@ -1,6 +1,6 @@
 import { RuleTree } from "./ruleTree";
 import { prettyPrintTree, Tree } from "../../util/tree";
-import { Rule, SingleCharRule, Span } from "./types";
+import { ParseError, Rule, SingleCharRule, Span } from "./types";
 
 // supposed to be like regex syntax
 export function prettyPrintCharRule(rule: SingleCharRule): string {
@@ -59,4 +59,12 @@ export function prettyPrintRuleTree(rt: RuleTree, source: string): string {
   return prettyPrintTree(ruleTreeToTree(rt, source), (n) =>
     renderRuleNode(n.item, source)
   );
+}
+
+export function prettyPrintParseError(error: ParseError): string {
+  return `offset ${error.offset}: expected ${
+    typeof error.expected === "string"
+      ? error.expected
+      : prettyPrintRule(error.expected)
+  }; got ${error.got}`;
 }
